requirements.txt
external==0.0.1
imageio==2.28.1
kornia==0.6.12
mathutils==2.81.2
matplotlib==3.5.2
numpy==1.21.6
open3d==0.16.0
opencv_contrib_python==4.6.0.66
opencv_python==4.11.0.86
pycuda==2022.1
pynvml==11.5.3
pyrender==0.1.45
PyYAML==6.0
scikit_image==0.19.3
scipy==1.7.3
torchmetrics==0.11.4
tqdm==4.65.0
trimesh==3.21.5
Pillow
pypose
git+https://github.com/facebookresearch/pytorch3d.git
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch